add newTab pixel, fire pixels only for new users first time (#1219)
Task/Issue URL:
https://app.asana.com/0/1204186595873227/1204669498383680/f

**Description**: Adds a pixel for the first time a user see the new tab
page. It also makes this and the activations pixel fire only when the
user is a new users (intended as within a week of the first time they
launch the app). Initially uses the cohort to determine if it’s the
first time the user has opened the app

Steps to test this PR:
1. ./clean-app.sh debug    
2. Run the app and check that the pixel "m_mac.new-tab-opened.initial”
is fired
3. close and run the app again
4. Open other tabs and check no pixel is fired
5. Enable email protection and check that the pixel
"m_mac.enable-email-protection.initial” is fired.
6. Disable and enable email protection again and check that the pixel is
not fired
7. Watch a video in Duck Player and check that the pixel
"m_mac.watch-in-duckplayer.initial” is fired.
8. Watch another video in Duck Player and check that the pixel is not
fired
9. Enable cookie pop up management and check that the pixel
"m_mac.cookie-management-enabled.initial” is fired.
10. Disable and enable cookie pop up management again and check that the
pixel is not fired
11. Make default browser and check that the pixel
"m_mac.set-as-default.initial” is fired.
12. Make another browser default and make DDG the default browser again
and check that the pixel is not fired
13. Import bookmarks and/or passwords and check that the pixel
“m_mac.import-data.initial” is fired
14. Import data again and check that the pixel is not fired
15. Kill and run the app again and check no pixel is fired for any of
the above actions.
16. /clean-app.sh debug    
17. In AppDelegate line 167 make == to != (or make atb return a value)
18. try all the actions again and check no pixel is sent

<!—
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
—>

—
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)