Find In Page bugs (#1241)
Task/Issue URL:
https://app.asana.com/0/1177771139624306/1204583183751067/f

**Description**:

**Steps to test this PR**:
Validate these test scenarios:
**Find field losing focus**

- CMD-F, search for a word on the page that is in an editable field
(e.g. asana task title)
- Esc and close the find UI
- CMD-F again...
- Bug: the find field loses keyboard focus, in favor of the item that
has been selected due to it being a match. The field that this word is
it steals focus.

**Re-open Find In Page**

- Open FindInpage, jump to second find result
- Hit Cmd+F again
- result: first match in page is highlighted
- expected: current (second) match is highlighted again

**Quickly type** 

- Open a page (e.g. https://ipsw.me/download/VirtualMac2,1/21G115)
- Quickly type "ma" - second match is highlighted
- expected: first match should be highlighted
- Close FindInPage
- Reopen FindInPage
- result: second match is highlighted
- expected: first match should be highlighted

**Re-type**

- Select all ("ma") in Find Field
- Type over "ma"
- result: next match is highlighted
- expected: current match should be re-highlighted
- or: first match in page highlighted if typed slowly enough and first
clear then type

**Navigation**

- Find In Page isn‘t closed on GH navigations (which are the same-doc
navigations)

**PDF**

- Open a PDF
- Search for something
- Hit cmd+g: same result is highlighted, but the match counter is
increased; After going round the ghost result is not highlighted (same
in Safari, but not in Preview.app)

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)