Update Xcode to 14.3.1 (#1274)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1204857106550931/f
Tech Design URL:
CC: @ayoy

Description:

This PR updates our Xcode version to 14.3.1. This version is used by the CI, so we will need this PR in order to deploy builds from with this version of Xcode. This is important as 14.3.1 has added some Swift Concurrency fixes for macOS versions 12.0 up to (but not including) 12.1.