Implements a mechanism to avoid processing status changes multiple times in the menu agent (#1243)
Task/Issue URL: https://app.asana.com/0/0/1204733709518353/f

## Description:

The agent app is sometimes sending out `.requestStatusUpdate`
Distributed Notifications as a way to know if the network extension and
connection are still alive. This results in the extension sending out
`.statusDidChange` notifications that may not always represent a status
change that just happened.

This PR introduces a mechanism that includes a timestamp in the
`statusDidChange` notification that the menu agent app checks to know if
it can safely ignore the status update.