Change how we save the encyption key (#1290)
Task/Issue URL:
https://app.asana.com/0/1177771139624306/1204466438263867/f

Description:
During migration to a new mac users lose their history because the
encryption key (used to encrypt and decrypt history and autofill data)
is not migrated into the new mac.
This PR changes the way the encryption key is saved. It used to be saved
in binary as an iCloud item with this change will be saved in bae64 and
as a login item. The change to login item should allow the key to be
migrated to the new mac. Saving the key in base 64 so that is visible
from keychain could be potentially use to fix problems after migration
afterwards.
If the key is not been saved in the new way yet it will read the
original one and save it in the new way