Improve NetP logging (#1259)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1204830533003048/f
Tech Design URL:
CC: @diegoreymendez

Description:

This PR enables NetP extension logging by default. It was disabled before the PR was merged, but I'm turning it back on for now as it doesn't get used unless NetP is actually running, and because of our disclosure terms we're able to perform such logging. In production, we wouldn't do this.

This PR also sets the subsystem of the NetP extension logs to com.duckduckgo.macos.browser.network-protection. Our current use of subsystem across the app isn't super useful - today, we just set it to the bundle ID, but it would be better to filter by subsystem in cases like this where we want to see all NetP logs at once.