1Password import instructions update (#1280)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1203939375506037/f
Tech Design URL:
CC:

Description:

This PR updates the 1Password import flow to include instructions for 1Password 7 and 1Password 8.

Note that some of the bolded text is deliberately missing - we kept it simple since this implementation would become overwhelmed with NSAttributedString logic if we followed the original designs. Later versions of macOS make this simple with Markdown support in strings, but for now we agreed to just ditch the bolded text in favour of a simple implementation.