Fix the missing VPN-16 image warning (#1321)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1204952080663215/f
Tech Design URL:
CC: @brindy @diegoreymendez

Description:

This PR fixes a warning related to one of the NetP icons.

This icon is set in the storyboard but doesn't exist in the app's asset catalog. It instead comes from a Swift package asset catalog and is set at runtime.