Make code_freeze resumable (#1287)
Task/Issue URL: https://app.asana.com/0/1203301625297703/1203860421711260/f

Description: 
As a part of the code_freeze flow, unit test are run after updating the embedded files to validate that
everything is ok before the release. If Unit Test fails, the flow is stopped and it's difficult for the
release manager to restart. This PR adds a resume option to the code_freeze lane.

When the option is set:
* Fastlane expects to run from a release/<something> branch, instead than develop.
* The repository is not expected to the clean (e.g. if unit test failed because of flakiness,
  the embedded files are updated, but not committed yet). A list of modified, not committed
   files is shown to the user when the flow restarts.