Update the check used to determine which user defaults wrapper is used (#1332)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1204964526109657/f
Tech Design URL:
CC: @ayoy

Description:

This PR fixes a test failure introduced recently in PR #1259.

This PR was using NETP_SYSTEM_EXTENSION as its check, but this value is actually set on the Developer ID target also. So, to fix this, we need to check for NETP_SYSTEM_EXTENSION and also NETWORK_EXTENSION, to ensure we're only targeting the system extension itself.