Adds the solving captcha action (#1333)
## Task

https://app.asana.com/0/1204006570077678/1204956531292317/f

## Description

Please take a look at the Asana task for acceptance criteria.

This pull request adds the `solveCaptcha` action. This action works with
a transaction ID from the previous `getCaptchaInfo` action.

With that transaction ID, we hit the backend, trying to fetch the
captcha-resolved data, then we pass that data to CCF to inject it into
the captcha.