update favicon logic to fix gsuite icons (#1380)
Task/Issue URL:
https://app.asana.com/0/1177771139624306/1203279312024256/f
Tech Design URL:
CC:

**Description**:

Updates favicon logic:

* Do not request /favicon.ico unless the list of icons returned by the
page is empty
* Expand the icon selector to include all icons
* Update the new tab page to use the full URL for the icons

**Steps to test this PR**:
1. Visit
https://docs.google.com/spreadsheets/d/1MFaoi0avVgJsr-u0IET88sM-Z7F4qvLXaZR5EKjaj4U/edit#gid=0
and add it as a favorite. It should have the green sheets icon in the
tab and on the new tab page.
2. Visit
https://docs.google.com/document/d/1VQELRLU_PyUCdZ8ATaMWAZqtzfHG6KZHksqPGFYkhzs/edit
and add it as a favorite. It should have the google docs blue/white icon
in the tab and on the new tab page.
3. Smoke test a bunch of sites and check the tab and favorites icon.
Choose some from near the top of this list and other more 'long tail'
like sites:
https://jupyter.duckduckgo.com/user/x/lab/tree/brindy/serp-top-sites.ipynb
- if any icons look noticeably difference compare with prod version of
the browser and see which is better. If this version looks worse, please
flag it.

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)