add in product survey to macOS (#1414)
Task/Issue URL:
https://app.asana.com/0/1204186595873227/1204886008020195/f

**Description**: Adds two new cards to the new tab: survey day0 and
survey day7. The first one is visible to user within a day of installing
the app the second after a week of installing the app. Clicking the card
takes you to the survey link. Users who clicked on the first card will
not be shown the day 7 card.
The all new tab cards section as well as specifically the survey cards
are behind a privacy config switch.
There is BSK update to add the new privacy config feature.