Changes external application prompt behaviour (#1437)
Task/Issue URL:
https://app.asana.com/0/1177771139624306/1205137373084365

**Description**: The external application prompt is transient (disappear
if users clicks elsewhere) and it does not appear if the user re-click
on the button that makes it appear. Users tends to dismiss it and then
thing the website is broken. To solve this we decided to make the prompt
non transient and make it appear at every click.