Turns NetworkProtectionStatusView into a composite view and separates the tunnel view. (#1466)
Task/Issue URL: https://app.asana.com/0/0/1205247622252423/f
Tech Design URL:
CC:

## Stack PR:

- Feature branch: `diego/feature-network-protection-onboarding`
    - https://github.com/duckduckgo/macos-browser/pull/1465
    - https://github.com/duckduckgo/macos-browser/pull/1466 (this PR)

## Description:

`NetworkProtectionStatusView` has grown to become quite hard to
maintain, and the same applies for the view's model.

I need to make some changes to be able to disable specifically the
tunnel controller part of the view, and these changes make it a bit
easier and cleaner.

This PR splits it into a parent composite view, and a child
`TunnelControllerView`.

## Scope:

This PR concentrates refactoring work, meaning that none of the changes
should affect functionality in any way.

The only intentional change has been to use the menu item attributes
that were missing. I'll mark this change with a comment.