Update input files checker for DBP (#1514)
Task/Issue URL: https://app.asana.com/0/0/1205318149426811/f
Tech Design URL:
CC: @Bunn @jotaemepereira @ayoy

Description:

This PR updates the input files checker to guarantee that the DBP target is consistently with the one it was based off of (the non-sandboxed target).

This PR also found that the DBP target was missing a couple source files, so those have been added.