Save/edit always the same profile (#1545)
## Asana

https://app.asana.com/0/1203581873609357/1205344732841496/f

## Description

The database schema supports multiple profiles, and because of that,
when saving a new profile, we insert new rows in the database. This is
causing some issues with testing.

When a new profile is saved, we replace the current one and **delete all
the current scan operations**. For example, if we save John Smith and
there are scan operations, and then we want to test with Ben Douglas,
all the scan operations of John Smith will be deleted to ensure we are
always scanning for the latest profile.

If opt-outs are in progress, we continue with those.