Add support for creating build variants to identify installation source (#1571)
Task/Issue URL: https://app.asana.com/0/72649045549333/1203384609277840/f

Description:
This change adds a workflow that is supposed to be run as the last step of the release process.
It takes a list of comma-separated ATB variants (two-letter combinations) and works on an already
released DMG to create build variants from it. A build variant is a DMG of the app with an additional
variant.txt file placed inside Resources directory of the bundle. Build variants are available
to download from https://staticcdn.duckduckgo.com/macos-desktop-browser/xy/duckduckgo.dmg,
where xy is the variant identifier.
If variant.txt is present in the bundle, the app appends it to the ATB value sent to exti endpoint
to attribute installation to variant's respective campaign.