Generalise login item manager (#1598)
Task/Issue URL: https://app.asana.com/0/0/1205424898479196/f
Tech Design URL:
CC:

**Description**:
Makes NetworkProtectionLoginItemsManager more generic so we can use it
for DBP.

In doing so, it also removes the state from it, it no longer keeps track
of LoginItems itself. I've got mixed feelings about this since an
individual LoginItem is a struct, but is always retrieved from a static
variable. If the last part changes, we could subtly break something. But
the three different preprocessor flags in play limit the options a
bit...

**Steps to test this PR**:
1. Test it compiles with every possible combination of the DBP,
NETWORK_PROTECTION, and NETP_SYSTEM_EXTENSION flags
2. Test NetP still works. This is where I have less knowledge of what to
test and how it might break. I've tried restarting my machine, and
disabling and reenabling NetP, and things seem to work

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)