NetP connection notifications improvements (#1730)
Task/Issue URL: https://app.asana.com/0/0/1205656863843386/f

**Description**:
On testing the Network Protection notifications on iOS, we decided to
make a few UX tweaks:
- Show the same connection notifications when the user connects manually
as well as when they reconnect after an interruption
- Include the server location in the body text of these notifications

**Steps to test this PR**:
1. Check out and build this branch
2. Reset the VPN extensions to bring in the new changes
3. Start NetP
4. **Observe the connection notification on connecting**
5. Go to Debug -> Network Protection -> Simulate Failure -> Connection
Interruption
6. **Observe the interruption notification followed by the connection
notification**

Note that, at least for system extension builds, the location doesn’t
seem to be available. This seems to be an existing bug that was out of
scope for these changes. There is, however, fallback copy for this case
which is the same as before, so I’m just keeping with that.

—
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)