Resets the context in HistoryStore after a failure to modify objects … (#1794)
Task/Issue URL:
https://app.asana.com/0/1199230911884351/1205851785198354/f

## Description

The need for this change was discussed here: [Comment by Diego Rey
Mendez on macOS App
Maintenance](https://app.asana.com/0/0/1205823163501929/1205842940740611/f)

We agreed to push for some changes so that we reset (context.reset())
the context used in HistoryStore on failure. This is meant to clean the
context after a failure so that we don't pollute future operations on
the same context. This should give us a cleaner picture to look at.