Make appcast version checker timezone aware (#1843)
Task/Issue URL:
https://app.asana.com/0/1203301625297703/1205913080437957/f
Tech Design URL:
CC:

**Description**:
Currently, the date in `.sparkle_tools_release_date` is parsed as a
local date, but the last modified date of the binary is UTC. This makes
the check unreliable.
This PR updates the checker to parse the date in
`.sparkle_tools_release_date` as a UTC date. The PR also updates the
current date in `.sparkle_tools_release_date` as the latest binaries
have been released on Oct, 15th at 10PM UTC.

P.S.: Note that the checker may issue false positive results in case of
multiple releases on the same day.

**Steps to test this PR**:
1. Run `appcastManager.swift` and verify that the check passes. 
2. Move forward the date in `.sparkle_tools_release_date`.
3. Run `appcastManager.swift` and verify that the check fails. 

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)