Address Bar Spoofing Remediations + Test Cases (#1815)
Task/Issue URL: https://app.asana.com/0/0/1205844442726343/f 
Tech Design URL: N/A
CC: N/A

**Description**:
The browser implementation of the address bar updating functionality is
currently vulnerable to several Address Bar Spoofing vulnerabilities.
All three of the current issues share a single root cause - we are
simply updating the address bar too soon in the navigation flow. See
[this task](https://app.asana.com/0/0/1205809497861072/f) to understand
how other browsers handle these situations. The aim of this PR is to
replicate the behaviour of other common browsers by updating the
`subscribeToUrl` function such that it doesn't update the address bar
until *after* the navigation - as opposed to *as soon as the navigation
starts*.

**Steps to test this PR**:
1. Make a search - see address bar update after search completes.
2. Click a search result - see address bar updates accordingly again.
3. Visit
https://privacy-test-pages.site/security/address-bar-spoofing/spoof-js-page-rewrite.html
4. Click *run* button
5. Ensure the address bar isn't spoofed (in the current version it will
be spoofed to display duckduckgo.com:8080)
6. Visit a page that doesn't exist to render the error page, i.e.
http://notreallyadomain.com/
7. Ensure the error page loads correctly and the address bar shows
"notreallyadomain.com"

My only challenge with this PR that I would like some validation on is
the removal of a call to `updateAddressBarStrings()` on line 59 of
TabViewModel inside errorViewState - in my testing I found that this
gets called on each tab load, and is always followed by a later call to
`updateAddressBarStrings()` inside `subscribeToUrl()` so it seemed
superfluous but I'm probably mistaken so I'd like to raise it here.

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)