Fix Address Bar Spoofing via Blob: URLs (#1891)
Task/Issue URL:
https://app.asana.com/0/1177771139624306/1205962067199910/f
More Context in Comments:
https://app.asana.com/0/1176047645786601/1205947928630574/f

**Description**:
Our browser is vulnerable to address bar spoofing by abusing the
unsupported basic auth credentials string used in blob URLs. I see no
reason why we should allow basic authentication strings in blob: URLs
since they aren't accepted by other browsers. As such we replicate the
behaviour of Firefox where we strip the basic auth part from the address
bar and append a GUID to the URL path. Note, WebKit already does this
for data: URLs so this PR also ensures consistency between how the
browser handles blob: and data: URLs.

**Steps to test this PR**:
1. Visit https://notarootkit.com/AddressBarSpoofing/blob.html
2. Click "exploit with blob"
3. Ensure address bar isn't spoofed
4. Go back
5. Click "exploit with data"
6. Ensure address bar isn't spoofed

The changes implemented here should not affect navigation or the use of
blob: URLs in the browser as a whole, only what we display if a webpage
attempts to rewrite the history to a blob: URL which is an edge case.

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)