DBP: Add scan profile pixels (#1919)
## Task
https://app.asana.com/0/1203581873609357/1205337273100855/f

## Description
Adds scan profile pixels. This pull request also fixes a bug where we
were swallowing WebView related errors when loading a URL.

When a timeout or other HTTP error codes happen, we catch those errors
and act accordingly. In the past, we treated them as non-errors, causing
a scan, for example, to be treated as non-matches found when an error
happened.