VPN Geoswitching - initial draft (#1978)
Task/Issue URL: https://app.asana.com/0/0/1206183910299715/f

Description:

This adds an initial implementation of location switching for the VPN. Known issues:

Some of the layout / spacing is not quite right
The tap area of the items on the VPN Location selection screen is too small
The styling of the list sections on the VPN Location screen is not finished.
There seems to be a bug where the initial selection is not respected when connecting after selecting
There is a momentary delay before the list items load.
Needs unit tests (I promise I will add them, they are very similar to the iOS ones so it won’t take long. I just ran out of time before the holidays).