Extend hover area for address bar's bookmark button over bookmarks bar and tab bar (#2020)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1203066850737780/f

Description:
This change removes MouseOverView from NavigationBarVC that was used to display bookmark button
in the address bar. It's replaced with tracking mouse events in MainView. If a mouse is moved in the area
inside the app window but above webView, the bookmark button is shown, otherwise it's hidden.
This way the bookmark button will also be shown when the cursor is over the bookmarks bar and over
the window title bar.