Resolve tunnel controller memory leak (#2087)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1206387864489596/f
Tech Design URL:
CC:

Description:

This PR fixes a retain cycle with the tunnel controller IPC client.

The issue is because the tunnel controller IPC client retains the XPC client, which retains the XPC connection, which retains the tunnel controller due to the exportedObject property.

The fix is to add a new object which the tunnel controller holds onto, and gives to the XPC connection as its exported object. That way, the XPC connection isn't retaining the IPC client, so it is free to get deallocated.