Remove admin user check.
Query creation is having serious problems when combined with `takeRetainedValue`. Using `takeUnretainedValue` fixes this issue, but leaks the query instances.

We don’t desperately need to know this value, so we’ll go without it and let the system extension API tell us when it doesn’t have admin privileges.