Simplifies some code (#2180)
Task/Issue URL:
https://app.asana.com/0/1199230911884351/1206556986192396/f

## Description

This PR brings balance back to the force. No I'm kidding - it fixes a
very strange issue in macOS 11.7 (x86) where some code was just being
swallowed up by the system (not crashing, and not proceeding through any
of the available flows).

This was probably "swallowed" due to `async / await`, but that's just a
guess.