NetP App Store waitlist tweaks (#2197)
Task/Issue URL: https://app.asana.com/0/1199230911884351/1206596097499805/f
Tech Design URL:
CC: @Bunn

Description:

This PR removes the call to action button for the NetP waitlist, and makes it US only. This is because we don't want tens of thousands of users to sign up - we'll make the App Store waitlist a little more understated.