Generate appcast2.xml and upload files to S3 in a CI workflow (#2184)
Task/Issue URL: https://app.asana.com/0/0/1206459737646671/f

Description:
This change adds new workflow called sparkle_internal.yml that should be triggered
manually after smoke testing. It takes Asana release task URL and the tag to be published,
and generates new appcast file and uploads missing files to S3.
The workflow fetches DMG for the release from S3 (based on the assumption that
Code Freeze and/or Bump Internal Release uploaded the DMG there – this was introduced
in #2165). appcastManager and upload_to_s3 scripts were updated for running in CI
(support non-interactive shell). Additionally, appcastManager generates a list of
added/updated files.
Completion is reported to the Automation subtask and also to the release task. Upon
successful completion, the release DRI is assigned a task to verify that the update worked,
with instructions on how to revert the appcast in case it's broken.
If the workflow fails, the user is assigned a task with instructions for generating appcast
and uploading to S3 locally.