Consolidate Letter Icon Creation in Autofill, Favorites and Recently Visited UI (#2201)
Task/Issue URL: https://app.asana.com/0/1177771139624306/1205337057550165/f

**Description**:
When we don’t have a cached Favicon, we display letters in place of the
Favicon. We currently do this in several places - the Autofill UI,
Favorites UI, and Recently Visited Sites. However, currently we lack
consistency in that two letters are displayed in Autofill, while one
letter is displayed in Favorites and Recently Visited Sites. This PR
aims to address this by consolidating the method used to display these
letters.